Fix the offline.js wipePostbackParamsThatAreNotForUs test
There was some funky module loading issue that was only happening in test.
That caused a subset of the urijs library to be pulled in.

The urijs pulled in looks like it came from insights-components
It differed in behavior, and it lacked some functions that should be present

This change should only affect tests