https://issues.redhat.com/browse/RHCLOUD-19050
There is a race condition between chrome history mutations and quickstarts mutations. We have no way of quickly fixing it in quickstarts so we will have to use poor man's solution. We should talk to the QS developers and ask them about adding a callback so we can handle the mutations ourselves (through our history instance).