Update fetch wrapper in iqeEnablement
The wrapper around fetch in iqeEnablement.ts was dropping headers if a
Request object was passed in as the input argument to fetch. This commit
ensures that if a Request object is passed that the headers are
preserved.

Note that this commit does NOT necessarily ensure full compatibility
with the fetch API... it simply takes one step in that direction by
making sure the headers are preserved.