Update fetch wrapper in iqeEnablement
Make the fetch wrapper compatible with the fetch API as it now exposes
the same interface as fetch is exposing.

To make it work, a Request is created with what is passed as a
parameter to the function. Request is taking the same parameters than
fetch which is quite handy in that case. Once we had a Request object
it can be upgraded with the custom header the way it was done before.