Update chr-scope css class
The height of .chr-scope__default-layout (which is a parent container
for Insights apps) is hardcoded as 100vh minus the height of the
masthead. The masthead heights are incorrect (something must have
changed) and this causes a scrollbar to appear in Insights apps, even
when the content should not overflow/have a scrollbar. This commit
updates the height for all viewport sizes.